# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Blog:
    properties:
      author_id:
        description: |-
          ID of the author who wrote the blog (required)
          example: 507f1f77bcf86cd799439022
        type: string
      author_name:
        description: |-
          Author's name (required)
          example: John Doe
        type: string
      content:
        description: |-
          Blog content body (required)
          example: This is an introductory post about the Go programming language...
        type: string
      created_at:
        description: |-
          Timestamp when blog was created
          example: 2025-08-10T15:04:05Z
        type: string
      id:
        description: |-
          Unique identifier for the blog (MongoDB ObjectID)
          example: 507f1f77bcf86cd799439011
        type: string
      stats:
        allOf:
        - $ref: '#/definitions/domain.BlogStats'
        description: Engagement metrics (views, likes, etc.)
      tags:
        description: |-
          List of tags associated with the blog post (required)
          example: ["programming", "golang", "tutorial"]
        items:
          type: string
        type: array
      title:
        description: |-
          Blog post title (required)
          example: Introduction to Go
        type: string
      updated_at:
        description: |-
          Timestamp when blog was last updated
          example: 2025-08-11T16:05:06Z
        type: string
    required:
    - author_id
    - author_name
    - content
    - tags
    - title
    type: object
  domain.BlogStats:
    properties:
      comments:
        description: |-
          Number of comments on the blog
          example: 45
        type: integer
      dislikes:
        description: |-
          Number of dislikes received
          example: 10
        type: integer
      likes:
        description: |-
          Number of likes received
          example: 256
        type: integer
      views:
        description: |-
          Number of times the blog was viewed
          example: 1234
        type: integer
    type: object
  domain.Comment:
    properties:
      blog_id:
        description: |-
          ID of the blog post this comment belongs to
          example: 507f1f77bcf86cd799439022
        type: string
      content:
        description: |-
          Content/text of the comment
          example: This is a great post! Thanks for sharing.
        type: string
      created_at:
        description: |-
          Timestamp when the comment was created
          example: 2023-08-10T15:04:05Z07:00
        type: string
      id:
        description: |-
          Unique identifier for the comment
          example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        description: |-
          Timestamp when the comment was last updated
          example: 2023-08-11T16:05:06Z07:00
        type: string
      user_id:
        description: |-
          ID of the user who wrote the comment
          example: 507f1f77bcf86cd799439033
        type: string
      username:
        description: |-
          Username of the commenter
          example: johndoe
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Blog API documentation.
  title: Blog API
  version: "1.0"
paths:
...
  /blogs:
    get:
      description: Get paginated and filtered list of blog posts
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Author name
        in: query
        name: author
        type: string
      - collectionFormat: csv
        description: Tag filter
        in: query
        items:
          type: string
        name: tag
        type: array
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List blog posts
      tags:
      - blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post by an authenticated user
      parameters:
      - description: Blog data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/domain.Blog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Blog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new blog post
      tags:
      - blogs
  /blogs/{blog_id}/comments:
    get:
      description: Retrieve comments with pagination for a specific blog post
      parameters:
      - description: Blog ID
        in: path
        name: blog_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of comments per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get paginated comments for a blog post
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a comment on a blog post. Requires authentication.
      parameters:
      - description: Blog ID
        in: path
        name: blog_id
        required: true
        type: string
      - description: Comment object
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/domain.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a new comment to a blog post
      tags:
      - Comments
  /blogs/{id}:
    delete:
      description: Delete a blog post if you are the author or an admin
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a blog post
      tags:
      - blogs
    get:
      description: Retrieve a single blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Blog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a blog post by ID
      tags:
      - blogs
    put:
      consumes:
      - application/json
      description: Update an existing blog post if you are the author
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated blog data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/domain.Blog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Blog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a blog post
      tags:
      - blogs
  /comments/{comment_id}:
    delete:
      description: Delete a comment if the user is the owner or admin
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a comment if the user is the owner
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: Updated comment content
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/domain.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing comment
      tags:
      - Comments
...
swagger: "2.0"